OCR figs w/ pathway_score > 0.5

+++ custom_pkgs/requirements.nix
@@ -1,11 +1,12 @@
-# generated using pypi2nix tool (version: 1.8.1)
-# See more at: https://github.com/garbas/pypi2nix
+# generated using pypi2nix tool (version: 2.0.0)
+# See more at: https://github.com/nix-community/pypi2nix
 #
 # COMMAND:
-#   pypi2nix -V 3 -e homoglyphs==1.3.1
+#   pypi2nix -V python37 -e homoglyphs==1.3.2
 #
 
-{ pkgs ? import <nixpkgs> {}
+{ pkgs ? import <nixpkgs> {},
+  overrides ? ({ pkgs, python }: self: super: {})
 }:
 
 let
@@ -17,16 +18,7 @@ let
   import "${toString pkgs.path}/pkgs/top-level/python-packages.nix" {
     inherit pkgs;
     inherit (pkgs) stdenv;
-    python = pkgs.python3;
-    # patching pip so it does not try to remove files when running nix-shell
-    overrides =
-      self: super: {
-        bootstrapped-pip = super.bootstrapped-pip.overrideDerivation (old: {
-          patchPhase = old.patchPhase + ''
-            sed -i               -e "s|paths_to_remove.remove(auto_confirm)|#paths_to_remove.remove(auto_confirm)|"                -e "s|self.uninstalled = paths_to_remove|#self.uninstalled = paths_to_remove|"                  $out/${pkgs.python35.sitePackages}/pip/req/req_install.py
-          '';
-        });
-      };
+    python = pkgs.python37;
   };
 
   commonBuildInputs = [];
@@ -35,16 +27,18 @@ let
   withPackages = pkgs':
     let
       pkgs = builtins.removeAttrs pkgs' ["__unfix__"];
-      interpreter = pythonPackages.buildPythonPackage {
-        name = "python3-interpreter";
-        buildInputs = [ makeWrapper ] ++ (builtins.attrValues pkgs);
+      interpreterWithPackages = selectPkgsFn: pythonPackages.buildPythonPackage {
+        name = "python37-interpreter";
+        buildInputs = [ makeWrapper ] ++ (selectPkgsFn pkgs);
         buildCommand = ''
           mkdir -p $out/bin
-          ln -s ${pythonPackages.python.interpreter}               $out/bin/${pythonPackages.python.executable}
-          for dep in ${builtins.concatStringsSep " "               (builtins.attrValues pkgs)}; do
+          ln -s ${pythonPackages.python.interpreter} \
+              $out/bin/${pythonPackages.python.executable}
+          for dep in ${builtins.concatStringsSep " "
+              (selectPkgsFn pkgs)}; do
             if [ -d "$dep/bin" ]; then
               for prog in "$dep/bin/"*; do
-                if [ -f $prog ]; then
+                if [ -x "$prog" ] && [ -f "$prog" ]; then
                   ln -s $prog $out/bin/`basename $prog`
                 fi
               done
@@ -55,18 +49,26 @@ let
           done
           pushd $out/bin
           ln -s ${pythonPackages.python.executable} python
-          ln -s ${pythonPackages.python.executable}               python3
+          ln -s ${pythonPackages.python.executable} \
+              python3
           popd
         '';
         passthru.interpreter = pythonPackages.python;
       };
+
+      interpreter = interpreterWithPackages builtins.attrValues;
     in {
       __old = pythonPackages;
       inherit interpreter;
-      mkDerivation = pythonPackages.buildPythonPackage;
+      inherit interpreterWithPackages;
+      mkDerivation = args: pythonPackages.buildPythonPackage (args // {
+        nativeBuildInputs = (args.nativeBuildInputs or []) ++ args.buildInputs;
+      });
       packages = pkgs;
       overrideDerivation = drv: f:
-        pythonPackages.buildPythonPackage (drv.drvAttrs // f drv.drvAttrs //                                            { meta = drv.meta; });
+        pythonPackages.buildPythonPackage (
+          drv.drvAttrs // f drv.drvAttrs // { meta = drv.meta; }
+        );
       withPackages = pkgs'':
         withPackages (pkgs // pkgs'');
     };
@@ -74,29 +76,33 @@ let
   python = withPackages {};
 
   generated = self: {
-
     "homoglyphs" = python.mkDerivation {
-      name = "homoglyphs-1.3.1";
-      src = pkgs.fetchurl { url = "https://files.pythonhosted.org/packages/0c/84/db578ac381fd59126d068d388e6a1c47d350d8f9822a9e58558a43def62f/homoglyphs-1.3.1.tar.gz"; sha256 = "b14a0d98eb2e050b2ad18b07a13ff01ba5773373913978152316a05a52e98277"; };
+      name = "homoglyphs-1.3.2";
+      src = pkgs.fetchurl {
+        url = "https://files.pythonhosted.org/packages/26/ba/3744c57befd5d2b23858c6cf0c420ef3f4294d5243509fa35966e08ad3f1/homoglyphs-1.3.2.tar.gz";
+        sha256 = "cdb61c1ed4d23a1ec297e3c9a483dba2de171164b98afc3ed43862a91605f2fc";
+};
       doCheck = commonDoCheck;
-      buildInputs = commonBuildInputs;
+      buildInputs = commonBuildInputs ++ [ ];
       propagatedBuildInputs = [ ];
       meta = with pkgs.stdenv.lib; {
         homepage = "https://github.com/orsinium/homoglyphs";
-        license = licenses.lgpl3Plus;
+        license = "GNU Lesser General Public License v3.0";
         description = "Get homoglyphs for text, convert text to ASCII.";
       };
     };
-
   };
   localOverridesFile = ./requirements_override.nix;
-  overrides = import localOverridesFile { inherit pkgs python; };
+  localOverrides = import localOverridesFile { inherit pkgs python; };
   commonOverrides = [
-
+    
+  ];
+  paramOverrides = [
+    (overrides { inherit pkgs python; })
   ];
   allOverrides =
     (if (builtins.pathExists localOverridesFile)
-     then [overrides] else [] ) ++ commonOverrides;
+     then [localOverrides] else [] ) ++ commonOverrides ++ paramOverrides;
 
 in python.withPackages
    (fix' (pkgs.lib.fold
